Removed RAM mirror and general refactoring
RAM mirror:

Space Invaders suffered from a memory corruption after approx 42 seconds after starting the game. This is approx where an invader flips the "Y".

Repro:

- Start the game.
- Leave it running for at least 42 seconds without any keyboard input.
- Insert coin and start the game after at least 42 seconds have elapsed.
- Notice that the game acts weird.

This seems to be related to the RAM mirror. By using the debugger (C++ debugger), we can observe that it occurs approx at the same time as the game enters the RAM mirror for the first time. The game seems to write to the ROM section of memory.

All RAM mirror functionality was removed, and it seems to work fine now.

Refactoring:

- Introduced typedefs for uint8_t, uint16_t, uint32_t and uint64_t.
- Shortening calls to functions with long namespaces by using "using".
- In instructions.h: Increase readability.